server/requirements.txt
flask
flask_restful
PyJWT
sqlalchemy
mysqlclient
flask_cors
